ADAPTSM-65 | @rebeccahongsf | update file structure and set interstitial page as default preview for content authoring
