+++ src/templates/storyblok-entry.js
@@ -51,14 +51,14 @@ class StoryblokEntry extends React.Component {
       content.component = 'interstitialPage';
     }
 
-    // Swap the page component for membership type of registrant interstitial page.
-    if (pageContext.membershipInterstitial) {
-      content.component = 'typeOfRegistrant';
-    }
-
     // Swap the page component for membership related contact interstitial page
     if (pageContext.membershipRelatedContact) {
       content.component = 'relatedContactSelection';
+    }
+
+    // Swap the page component for membership membership full payment form page.
+    if (pageContext.membershipFullPayment) {
+      content.component = 'membershipFullPaymentForm';
     }
 
     // Swap the page component for membership installments form page
